requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
anthropic==0.34.0
google-generativeai==0.8.0
redis==5.0.8
slowapi==0.1.9
python-dotenv==1.0.1
pydantic==2.8.0
secure==0.3.0
httpx<0.28
